GetArguments now uses getLine rather than '\n'
Changed the getArguments logic to use FileReader's getLine to get
everything up to the first newline, rather than re-create this logic in
the code.
This was prompted by the possibility that someone would open the keyfile
in notepad and then save it, inadvertently adding a carriage return to
the end of the last argument and thus making it impossible to properly
generate the key.